split slow tests into video file
