+++ tests/test_generate_video.py
@@ -23,189 +23,7 @@ IMAGE_2 = os.path.join(TEST_DATA_DIR,'prompt2.jpg')
 IMAGE_PROMPTS = f'{IMAGE_1}:0.5|{IMAGE_2}:0.5'
 TEST_VIDEO = os.path.join(TEST_DATA_DIR,'small.mp4')
 
-def test_image_png(testing_config, tmpdir):
-    '''Generate a single png image based on a text prompt
-    '''
-    config = testing_config
-    config.output_image_size = [128,128]
-    output_filename = str(tmpdir.mkdir('output').join('output.png'))
-    vqgan_clip.generate.image(eng_config=config,
-        text_prompts = 'A painting of flowers in the renaissance style:0.5|rembrandt:0.5^fish:0.2|love:1',
-        image_prompts = [],
-        noise_prompts = [],
-        init_image = [],
-        iterations = 5,
-        save_every = 50,
-        output_filename = output_filename)
-    assert os.path.exists(output_filename)
-    os.remove(output_filename)
-
-def test_image_bmp(testing_config, tmpdir):
-    '''Generate a single bmp image based on a text prompt. Testing a format with no package metadata support.
-    '''
-    config = testing_config
-    config.output_image_size = [128,128]
-    output_filename = str(tmpdir.mkdir('output').join('output.bmp'))
-    vqgan_clip.generate.image(eng_config=config,
-        text_prompts = 'A painting of flowers in the renaissance style:0.5|rembrandt:0.5^fish:0.2|love:1',
-        image_prompts = [],
-        noise_prompts = [],
-        init_image = [],
-        iterations = 5,
-        save_every = 50,
-        output_filename = output_filename)
-    assert os.path.exists(output_filename)
-    os.remove(output_filename)
-
-def test_image_jpg_save_every(testing_config, tmpdir):
-    '''Generate a single jpg image based on a text prompt
-    '''
-    config = testing_config
-    config.output_image_size = [128,128]
-    output_filename = str(tmpdir.mkdir('output').join('output.jpg'))
-    vqgan_clip.generate.image(eng_config=config,
-        text_prompts = 'A painting of flowers in the renaissance style:0.5|rembrandt:0.5^fish:0.2|love:1',
-        image_prompts = [],
-        noise_prompts = [],
-        init_image = [],
-        iterations = 5,
-        save_every = 2,
-        output_filename = output_filename)
-    assert os.path.exists(output_filename)
-    os.remove(output_filename)
-
-def test_image_jpg(testing_config, tmpdir):
-    '''Generate a single jpg image based on a text prompt, no save every
-    '''
-    config = testing_config
-    config.output_image_size = [128,128]
-    output_filename = str(tmpdir.mkdir('output').join('output.jpg'))
-    vqgan_clip.generate.image(eng_config=config,
-        text_prompts = 'A painting of flowers in the renaissance style:0.5|rembrandt:0.5^fish:0.2|love:1',
-        image_prompts = [],
-        noise_prompts = [],
-        init_image = [],
-        iterations = 5,
-        output_filename = output_filename)
-    assert os.path.exists(output_filename)
-    os.remove(output_filename)
-
-def test_image_story(testing_config, tmpdir):
-    '''Generate a single image based on a text prompt changing every 10 iterations
-    '''
-    config = testing_config
-    config.output_image_size = [128,128]
-    output_filename = str(tmpdir.mkdir('output').join('output.jpg'))
-    vqgan_clip.generate.image(eng_config=config,
-        text_prompts = 'A painting of flowers in the renaissance style:0.5|rembrandt:0.5^fish:0.2|love:1',
-        iterations = 100,
-        save_every = 50,
-        output_filename = output_filename)
-    assert os.path.exists(output_filename)
-    os.remove(output_filename)
-
-def test_image_noise_prompt(testing_config, tmpdir):
-    '''Generate a single image based on a noise prompt with save_every==50
-    '''
-    config = testing_config
-    config.output_image_size = [128,128]
-    output_filename = str(tmpdir.mkdir('output').join('output.jpg'))
-    vqgan_clip.generate.image(eng_config=config,
-        noise_prompts = '123:0.1|234:0.2|345:0.3',
-        iterations = 100,
-        save_every = 50,
-        output_filename = output_filename)
-    assert os.path.exists(output_filename)
-    os.remove(output_filename)
-
-def test_image_noise_prompt_story(testing_config, tmpdir):
-    '''Generate a single image based on a noise prompt no save every
-    '''
-    config = testing_config
-    config.output_image_size = [128,128]
-    output_filename = str(tmpdir.mkdir('output').join('output.jpg'))
-    vqgan_clip.generate.image(eng_config=config,
-        noise_prompts = '123:0.1|234:0.2|345:0.3^700',
-        iterations = 100,
-        output_filename = output_filename)
-    assert os.path.exists(output_filename)
-    os.remove(output_filename)
-
-def test_image_image_prompt(testing_config, tmpdir):
-    '''Generate a single image based on a image prompt
-    '''
-    config = testing_config
-    config.output_image_size = [128,128]
-    output_filename = str(tmpdir.mkdir('output').join('output.jpg'))
-    vqgan_clip.generate.image(eng_config=config,
-        image_prompts = IMAGE_PROMPTS,
-        iterations = 5,
-        output_filename = output_filename)
-    assert os.path.exists(output_filename)
-    os.remove(output_filename)
-
-def test_image_init_image(testing_config, tmpdir):
-    '''Generate a single image based on a image prompt
-    '''
-    config = testing_config
-    config.output_image_size = [128,128]
-    output_filename = str(tmpdir.mkdir('output').join('output.jpg'))
-    init_image = IMAGE_1
-    vqgan_clip.generate.image(eng_config=config,
-        text_prompts = 'A painting of flowers in the renaissance style',
-        init_image = init_image,
-        iterations = 5,
-        output_filename = output_filename)
-    assert os.path.exists(output_filename)
-    os.remove(output_filename)
-
-def test_image_all_prompts(testing_config, tmpdir):
-    '''Generate a single image based on a text prompt, image prompt, and noise prompt.
-    '''
-    config = testing_config
-    config.output_image_size = [128,128]
-    output_filename = str(tmpdir.mkdir('output').join('output.jpg'))
-    vqgan_clip.generate.image(eng_config=config,
-        text_prompts = 'A painting of flowers in the renaissance style:0.5|rembrandt:0.5^fish:0.2|love:1',
-        image_prompts = IMAGE_PROMPTS,
-        noise_prompts = '123:0.1|234:0.2|345:0.3^700',
-        iterations = 100,
-        output_filename = output_filename)
-    assert os.path.exists(output_filename)
-    os.remove(output_filename)
-
-def test_image_all_prompts_story(testing_config, tmpdir):
-    '''Generate a single image based on a text prompt, image prompt, and noise prompt.
-    '''
-    config = testing_config
-    config.output_image_size = [128,128]
-    output_filename = str(tmpdir.mkdir('output').join('output.jpg'))
-    vqgan_clip.generate.image(eng_config=config,
-        text_prompts = 'A painting of flowers in the renaissance style:0.5|rembrandt:0.5^fish:0.2|love:1',
-        image_prompts = IMAGE_PROMPTS,
-        noise_prompts = '123:0.1|234:0.2|345:0.3^700',
-        iterations = 100,
-        output_filename = output_filename)
-    assert os.path.exists(output_filename)
-    os.remove(output_filename)
-
-def test_image_no_folder(testing_config):
-    '''Output filename specified without a folder
-    '''
-    config = testing_config
-    config.output_image_size = [128,128]
-    output_filename = str('output.jpg')
-    vqgan_clip.generate.image(eng_config=config,
-        text_prompts = 'A painting of flowers in the renaissance style:0.5|rembrandt:0.5^fish:0.2|love:1',
-        image_prompts = [],
-        noise_prompts = [],
-        init_image = [],
-        iterations = 5,
-        save_every = 50,
-        output_filename = output_filename)
-    assert os.path.exists(output_filename)
-    os.remove(output_filename)
-
+@pytest.mark.slow
 def test_video(testing_config, tmpdir):
     '''Generate a zoom video based on a text prompt changing every 10 iterations
     '''
@@ -225,7 +43,7 @@ def test_video(testing_config, tmpdir):
         zoom_scale=1.0, 
         shift_x=0, 
         shift_y=0)
-    output_files = glob.glob(steps_path + os.sep + '*.png')
+    output_files = glob.glob(steps_path + os.sep + '*.jpg')
     assert len(output_files) == num_video_frames
 
     # test generating video
@@ -239,6 +57,7 @@ def test_video(testing_config, tmpdir):
         os.remove(f)
     os.remove(output_filename)
 
+@pytest.mark.slow
 def test_video_zoomed(testing_config, tmpdir):
     '''Generate a zoom video based on a text prompt changing every 10 iterations
     '''
@@ -258,7 +77,7 @@ def test_video_zoomed(testing_config, tmpdir):
         zoom_scale=1.02, 
         shift_x=1, 
         shift_y=1)
-    output_files = glob.glob(steps_path + os.sep + '*.png')
+    output_files = glob.glob(steps_path + os.sep + '*.jpg')
     assert len(output_files) == num_video_frames
 
     # test generating video
@@ -272,6 +91,7 @@ def test_video_zoomed(testing_config, tmpdir):
         os.remove(f)
     os.remove(output_filename)
 
+@pytest.mark.slow
 def test_video_smoothed(testing_config, tmpdir):
     '''Generate a zoom video based on a text prompt changing every 10 iterations
     '''
@@ -292,7 +112,7 @@ def test_video_smoothed(testing_config, tmpdir):
         shift_x=1, 
         shift_y=1,
         z_smoother=True)
-    output_files = glob.glob(steps_path + os.sep + '*.png')
+    output_files = glob.glob(steps_path + os.sep + '*.jpg')
     assert len(output_files) == num_video_frames
 
     # test generating video
@@ -307,6 +127,7 @@ def test_video_smoothed(testing_config, tmpdir):
         os.remove(f)
     os.remove(output_filename)
 
+@pytest.mark.slow
 def test_video_story_prompts(testing_config, tmpdir):
     '''Generate a zoom video based on a text prompt changing over time
     '''
@@ -328,7 +149,7 @@ def test_video_story_prompts(testing_config, tmpdir):
         shift_x=1, 
         shift_y=1,
         z_smoother=True)
-    output_files = glob.glob(steps_path + os.sep + '*.png')
+    output_files = glob.glob(steps_path + os.sep + '*.jpg')
     assert len(output_files) == num_video_frames
 
     # test generating video
@@ -343,6 +164,7 @@ def test_video_story_prompts(testing_config, tmpdir):
         os.remove(f)
     os.remove(output_filename)
 
+@pytest.mark.slow
 def test_video_all_prompts(testing_config, tmpdir):
     '''Generate a zoom video based on a text prompt changing every 10 iterations
     '''
@@ -379,6 +201,7 @@ def test_video_all_prompts(testing_config, tmpdir):
         os.remove(f)
     os.remove(output_filename)
 
+@pytest.mark.slow
 def test_style_transfer(testing_config, tmpdir):
     output_images_path = str(tmpdir.mkdir('video_frames'))
     original_video_frames = video_tools.extract_video_frames(TEST_VIDEO, 
@@ -395,7 +218,7 @@ def test_style_transfer(testing_config, tmpdir):
             current_source_frame_prompt_weight=0.1,
             current_source_frame_image_weight=0.1)
 
-    output_files = glob.glob(generated_video_frames_path + os.sep + '*.png')
+    output_files = glob.glob(generated_video_frames_path + os.sep + '*.jpg')
     assert len(output_files) > 0
     assert len(output_files) == len(original_video_frames)
     for f in output_files:
@@ -403,6 +226,7 @@ def test_style_transfer(testing_config, tmpdir):
     for f in original_video_frames:
         os.remove(f)
 
+@pytest.mark.slow
 def test_style_transfer_smoothed(testing_config, tmpdir):
     output_images_path = str(tmpdir.mkdir('video_frames'))
     original_video_frames = video_tools.extract_video_frames(TEST_VIDEO, 
@@ -420,7 +244,7 @@ def test_style_transfer_smoothed(testing_config, tmpdir):
             current_source_frame_image_weight=0.1,
             z_smoother=True)
 
-    output_files = glob.glob(generated_video_frames_path + os.sep + '*.png')
+    output_files = glob.glob(generated_video_frames_path + os.sep + '*.jpg')
     assert len(output_files) > 0
     assert len(output_files) == len(original_video_frames)
     for f in output_files:
